+++ rsm/generate_subject_map.py
@@ -56,7 +56,7 @@ def main():
     xml_tree = etree.fromstring(response)
 
     #XSL Transformation : transforms the person_index data
-    transform_xsl = proj_root + "bin/utils/person_index_transform.xsl"
+    transform_xsl = proj_root + "rsm/utils/person_index_transform.xsl"
     xslt = etree.parse(transform_xsl)
     transform = etree.XSLT(xslt)
     person_index_data = transform(xml_tree)
@@ -110,7 +110,7 @@ def main():
                 subjectmap_exceptions_root.append(exception_item)
 
     #Below code transforms the xml files to csv files
-    transform_xsl = proj_root + "bin/utils/xml2csv.xsl"
+    transform_xsl = proj_root + "rsm/utils/xml2csv.xsl"
     xslt = etree.parse(transform_xsl)
     transform = etree.XSLT(xslt)
