+++ rsm/utils/configuration_checker.py
