+++ rsm/utils/redcap_transactions.py
@@ -8,7 +8,7 @@ from lxml import etree
 file_dir = os.path.dirname(os.path.realpath(__file__))
 goal_dir = os.path.join(file_dir, "../../")
 proj_root = os.path.abspath(goal_dir)+'/'
-sys.path.insert(0, proj_root+'bin')
+sys.path.insert(0, proj_root+'rsm')
 
 
 class redcap_transactions:
